workspace/healthcare_rag/requirements.txt
sentence-transformers>=2.6
faiss-cpu>=1.7
numpy>=1.24
# For LLM-fallback (valgfritt):
# openai>=1.0
